refactor end point of getPodcastById to return Long

+++ app/src/main/java/com/shamardn/podcasttime/data/remote/ApiService.kt
@@ -14,6 +14,6 @@ interface ApiService {
 
     @GET("/lookup?entity=podcastEpisode")
     suspend fun getPodcastById(
-        @Query("id") trackId: Int,
+        @Query("id") trackId: Long,
     ): PodcastResponse<EpisodeDTO>
 }
